Add hasImg property
We can't put in a "default value" because we have no way to tell if there's
an image without making the request.  For the good of future image
organization, images must be stored in <full name>.jpg format - I don't want
to store the full string because it's just another thing to manage.